wordcloud-generator/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
cycler==0.11.0
easy-pil==0.1.9
fonttools==4.37.1
frozenlist==1.3.1
idna==3.3
kiwisolver==1.4.4
matplotlib==3.5.3
multidict==6.0.2
numpy==1.23.2
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.12
wordcloud==1.8.2.2
yarl==1.8.1
